+++ rclcpp_lifecycle/test/test_state_machine_info.cpp
@@ -34,40 +34,40 @@ TEST_F(TestStateMachineInfo, available_states) {
   auto test_node = std::make_shared<rclcpp_lifecycle::LifecycleNode>("testnode");
   std::vector<rclcpp_lifecycle::State> available_states =
     test_node->get_available_states();
-  EXPECT_EQ((unsigned int)11, available_states.size());
+  EXPECT_EQ(11, available_states.size());
 
   // Primary States
-  EXPECT_EQ((unsigned int)0, available_states[0].id());  // unknown
-  EXPECT_EQ((unsigned int)1, available_states[1].id());  // unconfigured
-  EXPECT_EQ((unsigned int)2, available_states[2].id());  // inactive
-  EXPECT_EQ((unsigned int)3, available_states[3].id());  // active
-  EXPECT_EQ((unsigned int)4, available_states[4].id());  // finalized
+  EXPECT_EQ(0, available_states[0].id());  // unknown
+  EXPECT_EQ(1, available_states[1].id());  // unconfigured
+  EXPECT_EQ(2, available_states[2].id());  // inactive
+  EXPECT_EQ(3, available_states[3].id());  // active
+  EXPECT_EQ(4, available_states[4].id());  // finalized
 
   // Transition States
-  EXPECT_EQ((unsigned int)10, available_states[5].id());  // configuring
-  EXPECT_EQ((unsigned int)11, available_states[6].id());  // cleaningup
-  EXPECT_EQ((unsigned int)12, available_states[7].id());  // shuttingdown
-  EXPECT_EQ((unsigned int)13, available_states[8].id());  // activating
-  EXPECT_EQ((unsigned int)14, available_states[9].id());  // deactivating
-  EXPECT_EQ((unsigned int)15, available_states[10].id());  // errorprocessing
+  EXPECT_EQ(10, available_states[5].id());  // configuring
+  EXPECT_EQ(11, available_states[6].id());  // cleaningup
+  EXPECT_EQ(12, available_states[7].id());  // shuttingdown
+  EXPECT_EQ(13, available_states[8].id());  // activating
+  EXPECT_EQ(14, available_states[9].id());  // deactivating
+  EXPECT_EQ(15, available_states[10].id());  // errorprocessing
 }
 
 TEST_F(TestStateMachineInfo, available_transitions) {
   auto test_node = std::make_shared<rclcpp_lifecycle::LifecycleNode>("testnode");
   std::vector<rclcpp_lifecycle::Transition> available_transitions =
     test_node->get_available_transitions();
-  EXPECT_EQ((unsigned int)25, available_transitions.size());
+  EXPECT_EQ(25, available_transitions.size());
   for (rclcpp_lifecycle::Transition & transition : available_transitions) {
     EXPECT_FALSE(transition.label().empty());
 
-    EXPECT_TRUE(transition.start_state().id() <= (unsigned int)4 ||
-      (transition.start_state().id() >= (unsigned int)10 &&
-      (transition.start_state().id() <= (unsigned int)15)));
+    EXPECT_TRUE(transition.start_state().id() <= 4 ||
+      (transition.start_state().id() >= 10 &&
+      (transition.start_state().id() <= 15)));
     EXPECT_FALSE(transition.start_state().label().empty());
 
-    EXPECT_TRUE(transition.goal_state().id() <= (unsigned int)4 ||
-      (transition.goal_state().id() >= (unsigned int)10 &&
-      (transition.goal_state().id() <= (unsigned int)15)));
+    EXPECT_TRUE(transition.goal_state().id() <= 4 ||
+      (transition.goal_state().id() >= 10 &&
+      (transition.goal_state().id() <= 15)));
     EXPECT_FALSE(transition.goal_state().label().empty());
   }
 }
